git-info: temporarily set IFS to $'\t' for...
extracting ahead and behind from the output of

    git rev-list --count --left-right

The problem with the current implementation is that it is not
temper-proof against a modified IFS. The output format is
count1<TAB>count2, and ahead and behind are parsed through word
splitting, but if IFS doesn't contain the TAB character, then splitting
will fail, leading to bad math expression.

In this commit, we temporarily set IFS to $'\t', then restore it when
done.