zmwangx theme: now capable of update during line continuation
Two new parameters: prompt_zmwangx_accumulated_buffer_lines and
prompt_zmwangx_cursor_line, where the former is the total number of
accepted buffer lines, and the latter is the line in the current buffer
where the cursor is located.

Also in this commit: a newline is appended after the output of each
command.

See also: https://github.com/sindresorhus/pure/issues/143.