add autosuggestions module
But not using it since there are problems; e.g., tab completion is
dumbed down in many ways (e.g., cd completes to cd commands in history,
not to subdirectories in the current directory; even when I toggle it
off, there are still glitches, e.g., highlighting leftovers).

Not to mention it works extremely badly with user-customized tab (^I)
keybinding. I'm already monkey patching autosuggest-tab here, but still,
there can be circular dependencies where the original binding calls
expand-or-complete, but zsh-autosuggestions rebinds expand-or-complete
to autosuggest-tab.