zmwangx prompt: prevent locked repo when fetching/pulling
The async fetch job done by the prompt could lock the repo and result in
an error when user is trying to fetch or pull interactively. In this
commit, we flush the jobs when the command line is git pull or git
fetch.

Credit to Mathias Fredriksson.

See https://github.com/sindresorhus/pure/issues/162 and
commit 946edbf3377acb8ac82848aa690ac82b019b055d (http://git.io/vlCu3).