transmission/init.zsh: Improve torrent spec in messages
Print all spec in comma-and-space-separated form. If a single-word spec
is comma-separated, insert spaces too for consistency.

In fact, at this point, since we are no longer validating arguments,
it's kind of useless to have a central dispatcher. But in the spirit of
if it ain't broke don't fix it, let's just keep it.