acdcli: Redesign global options passing mechanism to external actions
The problem is that one can easily store some global options in an
alias (e.g., my acdcli is aliased to `acdcli --check=none'), then with
the old mechanism, all global options, including ones that are
unintentionally added, are passed to external actions as positional
arguments, which then need to be parsed alongside local options that
really matter. This has to be done in every external action, which is a
chore, and worse still, not future-proof, since the traditional option
parsing code that involves getopt or getopts will be screwed if a new
global option is added in the future. For the reasons above, separation
of global options and local ones is necessary, and obviously we're left
with one option: environment variable.

The new mechanism is thoroughly documented in modules/acdcli/README.md
and won't be repeated here (the reasoning is in a footnote).