python/functions/virtual-install: Absolutify project root...
and do not rely on pushd & popd.

pushd & popd doesn't work when the destination is ., in which case pushd
does nothing, and popd pops to the last directory on the stack, which
screws up all assumptions. Let's just do it the dump yet unsurprising
way: save old pwd in a variable, then cd back to old pwd when done.