zmwangx prompt: Improve robustness against terminal resizing
Previously, due to the promptsp option instructing the terminal to fill
the current line with spaces until wrap then print a CR (controlled by
promptcr), the preprompt line has a lot of pointless spaces that will
create a mess when the terminal is horizontally shrinked.

This commit clears space-only lines before printing the preprompt. A
genuinely partial line that is preserved by promptsp will still have a
lot of spaces though, and there's nothing I can do about it. (I've tried
the promptsp function shipped with zsh that is supposed to only print a
newline when the last line is truly partial, but like all approaches
involving reading the actual cursor position on the terminal screen, it
works badly — noticible delay, and utterly broken when screen is
cleared.)