Create new module gpg2
And move gpg/functions to gpg2.

The gpg module is used to set up the agent, and is really meant for gpg
v1. gpg2 always requires the agent, so there's no need for agent setup.
Nevertheless, I don't want to remove that original module, hence this
new one which only provides an alias (gpg -> gpg2) at this moment, as
well as my old gpgauth function.